[auth][#6] correct the token creation flow

+++ src/Authorization/OAuth20/Projects/DotNetExtensions.Authorization.OAuth20.Server/Domain/Client.cs
@@ -50,7 +50,7 @@ public class Client : EntityBase<int>
 
     public TokenType? TokenType { get; set; }
 
-    public long? TokenExpirationSeconds { get; set; }
+    public int? TokenExpirationSeconds { get; set; }
 
     public IEnumerable<ClientScope>? ClientScopes { get; set; }
